Bugfix for Linear RENs and Type Stability

+++ test/Wrappers/diff_ren.jl
@@ -13,7 +13,7 @@ Test that backpropagation runs and parameters change
 batches = 10
 nu, nx, nv, ny = 4, 5, 10, 3
 γ = 10
-ren_ps = LipschitzRENParams{Float64}(nu, nx, nv, ny, γ)
+ren_ps = LipschitzRENParams{Float32}(nu, nx, nv, ny, γ)
 model = DiffREN(ren_ps)
 
 # Dummy data
